+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `jafgen` now takes a `--prefix` argument that appends a prefix with an underscore to the output file names.
+
 ## [v0.2.1](https://github.com/dbt-labs/jaffle-shop-generator/releases/tag/v0.2.1) - 2023-01-06
 
 ## [v0.2.0](https://github.com/dbt-labs/jaffle-shop-generator/releases/tag/v0.2.0) - 2023-01-06